requirements.txt
SQLAlchemy>=1.4.15
snowflake-connector-python>=2.8.1
snowflake-sqlalchemy>=1.2.4
pandas>=1.2.4,<2.1